feat: extract bincode function to pkg/bincode
