+++ pkg/bincode/serialize.go
@@ -1,4 +1,4 @@
-package common
+package bincode
 
 import (
 	"encoding/binary"
