+++ pkg/bincode/varbyte_test.go
@@ -1,4 +1,4 @@
-package common
+package bincode
 
 import (
 	"reflect"
