+++ program/system/instruction.go
@@ -2,6 +2,7 @@ package sysprog
 
 import (
 	"github.com/portto/solana-go-sdk/common"
+	"github.com/portto/solana-go-sdk/pkg/bincode"
 	"github.com/portto/solana-go-sdk/types"
 )
 
@@ -23,7 +24,7 @@ const (
 )
 
 func CreateAccount(fromAccount, newAccount, owner common.PublicKey, initLamports, accountSpace uint64) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Lamports    uint64
 		Space       uint64
@@ -49,7 +50,7 @@ func CreateAccount(fromAccount, newAccount, owner common.PublicKey, initLamports
 }
 
 func Assign(accountPubkey, assignToProgramID common.PublicKey) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction       Instruction
 		AssignToProgramID common.PublicKey
 	}{
@@ -70,7 +71,7 @@ func Assign(accountPubkey, assignToProgramID common.PublicKey) types.Instruction
 }
 
 func Transfer(from, to common.PublicKey, lamports uint64) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Lamports    uint64
 	}{
@@ -92,7 +93,7 @@ func Transfer(from, to common.PublicKey, lamports uint64) types.Instruction {
 }
 
 func CreateAccountWithSeed(fromPubkey, newAccountPubkey, basePubkey, programID common.PublicKey, seed string, lamports, space uint64) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Base        common.PublicKey
 		Seed        string
@@ -128,7 +129,7 @@ func CreateAccountWithSeed(fromPubkey, newAccountPubkey, basePubkey, programID c
 }
 
 func AdvanceNonceAccount(noncePubkey, authPubkey common.PublicKey) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 	}{
 		Instruction: InstructionAdvanceNonceAccount,
@@ -149,7 +150,7 @@ func AdvanceNonceAccount(noncePubkey, authPubkey common.PublicKey) types.Instruc
 }
 
 func WithdrawNonceAccount(noncePubkey, authPubkey, toPubkey common.PublicKey, lamports uint64) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Lamports    uint64
 	}{
@@ -174,7 +175,7 @@ func WithdrawNonceAccount(noncePubkey, authPubkey, toPubkey common.PublicKey, la
 }
 
 func InitializeNonceAccount(noncePubkey, authPubkey common.PublicKey) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Auth        common.PublicKey
 	}{
@@ -197,7 +198,7 @@ func InitializeNonceAccount(noncePubkey, authPubkey common.PublicKey) types.Inst
 }
 
 func AuthorizeNonceAccount(noncePubkey, oriAuthPubkey, newAuthPubkey common.PublicKey) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Auth        common.PublicKey
 	}{
@@ -219,7 +220,7 @@ func AuthorizeNonceAccount(noncePubkey, oriAuthPubkey, newAuthPubkey common.Publ
 }
 
 func Allocate(accountPubkey common.PublicKey, space uint64) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Space       uint64
 	}{
@@ -240,7 +241,7 @@ func Allocate(accountPubkey common.PublicKey, space uint64) types.Instruction {
 }
 
 func AllocateWithSeed(accountPubkey, basePubkey, programID common.PublicKey, seed string, space uint64) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Base        common.PublicKey
 		Seed        string
@@ -267,7 +268,7 @@ func AllocateWithSeed(accountPubkey, basePubkey, programID common.PublicKey, see
 	}
 }
 func AssignWithSeed(accountPubkey, assignToProgramID, basePubkey common.PublicKey, seed string) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction       Instruction
 		Base              common.PublicKey
 		Seed              string
@@ -293,7 +294,7 @@ func AssignWithSeed(accountPubkey, assignToProgramID, basePubkey common.PublicKe
 }
 
 func TransferWithSeed(from, to, base, programID common.PublicKey, seed string, lamports uint64) types.Instruction {
-	data, err := common.SerializeData(struct {
+	data, err := bincode.SerializeData(struct {
 		Instruction Instruction
 		Lamports    uint64
 		Seed        string
